gleam-skin-main/backend/requirements.txt
django
djangorestframework
django-cors-headers
huggingface_hub
python-dotenv
mongoengine
bcrypt
